refactor: extract MIME→plugin mapping to shared utility
Move pluginForMime() from iframe-libraries.ts to a new shared module
at src/lib/renderer-plugins.ts. Add computeRequiredPlugins() for
scanning Jupyter outputs and returning required plugin names.

This module has no React/Vite/iframe dependencies — safe to import
from the materialization layer, which will use it to pre-compute
required plugins before React renders.
